pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wq-bus"
version = "0.1.0"
description = "WorldQuant BRAIN alpha research bus-architecture"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "pyyaml>=6.0",
    "click>=8.1",
    "aiohttp>=3.9",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
crawler = ["playwright>=1.40", "pymupdf>=1.23", "beautifulsoup4>=4.12"]
analysis = ["numpy>=1.26", "pandas>=2.1", "scipy>=1.11"]
dev = ["pytest>=7.4", "pytest-asyncio>=0.23"]

[project.scripts]
wqbus = "wq_bus.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["wq_bus*"]
